Export HashMap through dedicated API header
