+++ src/test/hashmap_concurrency_test.cpp
@@ -16,7 +16,7 @@
 #include <vector>
 
 #include "../error.h"
-#include "../hash_map.h"
+#include "../hash_map_api.h"
 #include "../mabain_consts.h"
 #include "../resource_pool.h"
